refactor(orchestration): move attention aggregation to ops, scope blocked-run query, harden unread (#4592)
- Relocate command_center_needs_input + gather_unread_signals (and the unread
  test) from schemas.rs into ops.rs, keeping schemas.rs a thin RPC wrapper per
  the canonical module shape.
- Query the ledger filtered to AwaitingUser runs so the run limit bounds
  blocked runs only; the prior global-page-then-filter could page an older
  still-blocked run out in a busy workspace.
- Make the unread-store read best-effort (log + empty) like the command-center
  read, so a transient local-DB hiccup can't sink already-resolved signals.
