+++ src/openhuman/orchestration/ops.rs
@@ -916,12 +916,135 @@ pub(crate) fn build_self_identity(
     }
 }
 
+// ── Attention queue aggregation ─────────────────────────────────────────────
+//
+// The `orchestration_attention` handler in [`super::schemas`] awaits the async
+// approval gate itself, then delegates the two synchronous source reads below.
+// Both are best-effort: a source failure degrades to an empty bucket (logged)
+// so the surviving signals still surface. The neutral-signal → item mapping is
+// the pure, unit-tested code in [`super::attention`].
+
+/// Cap on the command-center runs scanned for the `NeedsInput` bucket — the
+/// attention zone only needs the currently-blocked runs, not the full ledger.
+const ATTENTION_RUN_LIMIT: u32 = 100;
+
+/// Fetch the command-center `NeedsInput` bucket as neutral attention signals.
+/// Best-effort — a read error yields an empty vec (logged) so the rest of the
+/// attention queue still assembles.
+///
+/// The ledger query is filtered to `AwaitingUser` runs so [`ATTENTION_RUN_LIMIT`]
+/// bounds *blocked* runs only. Fetching a global recent page then filtering (as
+/// `list_agent_work` does) would let an older still-blocked run be paged out by
+/// newer working/completed runs in a busy workspace, silently dropping it from
+/// the attention queue.
+pub(super) fn command_center_needs_input(
+    config: &Config,
+) -> Vec<super::attention::NeedsInputSignal> {
+    use crate::openhuman::agent_orchestration::command_center::build_view;
+    use crate::openhuman::session_db::run_ledger::{
+        list_agent_runs, AgentRunListRequest, AgentRunStatus,
+    };
+    let request = AgentRunListRequest {
+        status: Some(AgentRunStatus::AwaitingUser.as_str().to_string()),
+        kind: None,
+        parent_run_id: None,
+        parent_thread_id: None,
+        limit: Some(ATTENTION_RUN_LIMIT),
+        offset: None,
+    };
+    match list_agent_runs(config, &request) {
+        Ok(response) => {
+            super::attention::needs_input_from_command_center(build_view(response.runs))
+        }
+        Err(e) => {
+            log::warn!(target: LOG, "[orchestration_rpc] attention.command_center_failed: {e}");
+            Vec::new()
+        }
+    }
+}
+
+/// Gather unread attention signals from the orchestration store: every non-pinned
+/// session with a positive unread count. The pinned master/subconscious windows
+/// are excluded — they are not agent instances.
+pub(super) fn gather_unread_signals(
+    conn: &rusqlite::Connection,
+) -> anyhow::Result<Vec<super::attention::UnreadSignal>> {
+    let mut out: Vec<super::attention::UnreadSignal> = Vec::new();
+    for session in store::list_sessions(conn)? {
+        if matches!(session.session_id.as_str(), "master" | "subconscious") {
+            continue;
+        }
+        let unread = store::unread_count(conn, &session.session_id)?;
+        if unread > 0 {
+            out.push(super::attention::UnreadSignal {
+                session_id: session.session_id,
+                label: session.label,
+                unread,
+                last_message_at: Some(session.last_message_at),
+            });
+        }
+    }
+    Ok(out)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::openhuman::orchestration::types::OrchestrationMessage;
     use std::sync::atomic::{AtomicUsize, Ordering};
     use tinyagents::graph::checkpoint::Checkpointer;
+
+    #[test]
+    fn gather_unread_signals_skips_pinned_and_zero_unread() {
+        use super::super::types::ChatKind;
+        let tmp = tempfile::tempdir().unwrap();
+        let config = Config {
+            workspace_dir: tmp.path().to_path_buf(),
+            ..Config::default()
+        };
+        let sess = |id: &str, source: &str, label: Option<&str>, at: &str| OrchestrationSession {
+            session_id: id.into(),
+            agent_id: "@peer".into(),
+            source: source.into(),
+            label: label.map(str::to_string),
+            workspace: None,
+            last_seq: 1,
+            created_at: "2026-07-06T00:00:00Z".into(),
+            last_message_at: at.into(),
+        };
+        let message = |id: &str, session: &str, kind: ChatKind, at: &str| OrchestrationMessage {
+            id: id.into(),
+            agent_id: "@peer".into(),
+            session_id: session.into(),
+            chat_kind: kind,
+            role: "user".into(),
+            body: "hello".into(),
+            timestamp: at.into(),
+            seq: 1,
+        };
+
+        let signals = store::with_connection(&config.workspace_dir, |conn| {
+            // Non-pinned session with one unread message → surfaces.
+            store::upsert_session(conn, &sess("h-1", "claude", Some("Claude · audit"), "t1"))?;
+            store::insert_message(conn, &message("m1", "h-1", ChatKind::Session, "t1"))?;
+            // Pinned master with a message → excluded (not an agent instance).
+            store::upsert_session(conn, &sess("master", "core", None, "t2"))?;
+            store::insert_message(conn, &message("m2", "master", ChatKind::Master, "t2"))?;
+            // Non-pinned session with no messages → zero unread, dropped.
+            store::upsert_session(conn, &sess("h-quiet", "codex", None, "t0"))?;
+            gather_unread_signals(conn)
+        })
+        .unwrap();
+
+        assert_eq!(
+            signals.len(),
+            1,
+            "only the non-pinned unread session surfaces"
+        );
+        assert_eq!(signals[0].session_id, "h-1");
+        assert_eq!(signals[0].unread, 1);
+        assert_eq!(signals[0].label.as_deref(), Some("Claude · audit"));
+    }
 
     #[test]
     fn self_identity_marks_published_identity_discoverable() {
